Small fixes

+++ auxiliar.h
@@ -2,7 +2,7 @@
 #define AUX_H
 
 void setMotors(int motor_left, int motor_right);
-int readUS();
+long readUS();
 
 
 #endif
